added authorization to endpoints

+++ ClubApi/ClubApi/Controllers/EventController.cs
@@ -3,13 +3,16 @@ using Application.Models;
 using Application.Models.Request;
 using Application.Services;
 using Domain.Entities;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace ClubApi.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class EventController : ControllerBase
     {
         private readonly IEventService _eventService;
@@ -18,6 +21,7 @@ namespace ClubApi.Controllers
             _eventService = eventService;
         }
 
+        [AllowAnonymous]
         [HttpGet("Events")]
         public ActionResult<ICollection<EventDto>> GetAllEvents()
         {
@@ -32,6 +36,7 @@ namespace ClubApi.Controllers
             }
         }
 
+        [AllowAnonymous]
         [HttpGet("ByName/{name}")]
         public ActionResult<EventDto> GetEventByName(string name)
         {
@@ -55,6 +60,10 @@ namespace ClubApi.Controllers
         {
             try
             {
+                var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (userRole == null || userRole == "Member")
+                    return Forbid();
+
                 return Ok(_eventService.GetMembersFromEvent(id));
             }
             catch (Exception ex)
@@ -69,6 +78,10 @@ namespace ClubApi.Controllers
         {
             try
             {
+                var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (userRole != "Admin")
+                    return Forbid();
+
                 return Ok(_eventService.CreateEvent(eventDto, creatorName));
             }
             catch (Exception ex)
@@ -83,6 +96,10 @@ namespace ClubApi.Controllers
         {
             try
             {
+                var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (userRole != "Admin")
+                    return Forbid();
+
                 _eventService.UpdateEvent(id, eventDto);
                 return NoContent();
             }
@@ -98,6 +115,10 @@ namespace ClubApi.Controllers
         {
             try
             {
+                var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (userRole != "Admin")
+                    return Forbid();
+
                 _eventService.DeleteEvent(id);
                 return Ok();
             }
@@ -112,6 +133,10 @@ namespace ClubApi.Controllers
         {
             try
             {
+                var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (userRole != "Member")
+                    return Forbid();
+
                 _eventService.SignUpEvent(memberId, eventId);
                 return Ok("El miembro se ha inscipto correctamente");
             }
@@ -125,6 +150,10 @@ namespace ClubApi.Controllers
         {
             try
             {
+                var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (userRole != "Member")
+                    return Forbid();
+
                 _eventService.DropOutEvent(memberId, eventId);
                 return Ok("El miembro se desinscribió correctamente");
             }
@@ -138,6 +167,10 @@ namespace ClubApi.Controllers
         {
             try
             {
+                var userRole = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (userRole != "Director")
+                    return Forbid();
+
                 _eventService.ApproveEvent(directorId, eventId);
                 return Ok("El evento fue aprobado correctamente");
             }
